Change double type to float

+++ backend/PhotovoltaicSystemCalculation/PhotovoltaicSystemCalculation/ExternalAPI/Models/SolarDTO.cs
@@ -3,6 +3,6 @@
     public class SolarDTO
     {
         public int Year { get; set; }
-        public IDictionary<int, double> Irradiance { get; set; } // key = month, value = year
+        public IDictionary<int, float> Irradiance { get; set; } // key = month, value = year
     }
 }
