+++ backend/PhotovoltaicSystemCalculation/PhotovoltaicSystemCalculation/Repositories/Models/WeatherDTO.cs
@@ -3,7 +3,7 @@
     public class WeatherDTO
     {
         public DateTime Date { get; set; }
-        public double CloudCover { get; set; }
-        public double Temperature { get; set; } // Kelvin
+        public float CloudCover { get; set; }
+        public float Temperature { get; set; } // Kelvin
     }
 }
